python-agents/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langgraph==0.0.55
neo4j==5.20.0
kafka-python==2.0.2
pytest==8.2.0
pydantic==2.7.1
python-json-logger==2.0.7
